Implement stats for containerd
This is a single endpoint that clients will poll
