+++ supervisor/event.go
@@ -23,8 +23,6 @@ const (
 	CreateCheckpointEventType EventType = "createCheckpoint"
 	DeleteCheckpointEventType EventType = "deleteCheckpoint"
 	StatsEventType            EventType = "events"
-	UnsubscribeStatsEventType EventType = "unsubscribeStats"
-	StopStatsEventType        EventType = "stopStats"
 	OOMEventType              EventType = "oom"
 )
 
@@ -59,7 +57,7 @@ type Event struct {
 	Checkpoint    *runtime.Checkpoint
 	Err           chan error
 	StartResponse chan StartResponse
-	Stats         chan interface{}
+	Stat          chan *runtime.Stat
 	CloseStdin    bool
 	ResizeTty     bool
 	Width         int
