+++ supervisor/metrics.go
@@ -6,6 +6,7 @@ var (
 	ContainerCreateTimer   = metrics.NewTimer()
 	ContainerDeleteTimer   = metrics.NewTimer()
 	ContainerStartTimer    = metrics.NewTimer()
+	ContainerStatsTimer    = metrics.NewTimer()
 	ContainersCounter      = metrics.NewCounter()
 	EventSubscriberCounter = metrics.NewCounter()
 	EventsCounter          = metrics.NewCounter()
@@ -19,6 +20,7 @@ func Metrics() map[string]interface{} {
 		"container-create-time": ContainerCreateTimer,
 		"container-delete-time": ContainerDeleteTimer,
 		"container-start-time":  ContainerStartTimer,
+		"container-stats-time":  ContainerStatsTimer,
 		"containers":            ContainersCounter,
 		"event-subscribers":     EventSubscriberCounter,
 		"events":                EventsCounter,
